Refactor - separación de algunos métodos en service
