+++ src/fixture/view/Ventana.java
@@ -6077,280 +6077,6 @@ public class Ventana extends javax.swing.JFrame {
         leerGolesDeEliminacionYGuardarCambios(Fase.FINAL);
     }//GEN-LAST:event_guardarBtnFinalActionPerformed
 
-    private ArrayList<Equipo> ordenarEquiposYCompletarOctavos(HashSet<Equipo> e, Grupo grupo, boolean guardandoCambios) {
-
-        ArrayList<Equipo> equipos = new ArrayList(e);
-
-        // Criterios de ordenamiento:
-        //  1) Mayor puntaje
-        //  2) Mayor diferencia de goles
-        //  3) Mayor cantidad de goles totales
-        // Dispongo de dos booleanos para informar al usuario si se requiere selección manual
-        boolean primerPuestoPendiente = false;
-        boolean segundoPuestoPendiente = false;
-
-        // Los ordeno por puntaje
-        Collections.sort(equipos, new Comparator<Equipo>() {
-            @Override
-            public int compare(Equipo e1, Equipo e2) {
-                return e1.getPuntos() > e2.getPuntos() ? -1 : 1;
-            }
-        });
-
-        // Verifico si es necesario reordenar por 
-        boolean puntajesIguales = false;
-        int auxPuntaje = -1;
-        for (Equipo equipo : equipos) {
-            if (equipo.getPuntos() == auxPuntaje) {
-                puntajesIguales = true;
-            }
-            auxPuntaje = equipo.getPuntos();
-        }
-
-        // Los reordeno por diferencia de goles si encontré dos equipos con igual puntaje
-        if (puntajesIguales) {
-            Collections.sort(equipos, new Comparator<Equipo>() {
-                @Override
-                public int compare(Equipo e1, Equipo e2) {
-                    int diferenciaDeGolesE1 = e1.getGolesHechos() - e1.getGolesEnContra();
-                    int diferenciaDeGolesE2 = e2.getGolesHechos() - e2.getGolesEnContra();
-
-                    if (e1.getPuntos() == e2.getPuntos()) {
-                        return diferenciaDeGolesE1 > diferenciaDeGolesE2 ? -1 : 1;
-                    }
-                    return 0;
-                }
-            });
-
-            // Verifico is es necesario reordenar por golesHechos
-            boolean diferenciaDeGolesIguales = false;
-            int auxDiferenciaDeGoles = equipos.get(0).getGolesHechos() - equipos.get(0).getGolesEnContra();
-            for (int i = 1; i < equipos.size(); i++) {
-                int diferenciaDeGoles = equipos.get(i).getGolesHechos() - equipos.get(i).getGolesEnContra();
-                if (diferenciaDeGoles == auxDiferenciaDeGoles) {
-                    diferenciaDeGolesIguales = true;
-                }
-                auxDiferenciaDeGoles = diferenciaDeGoles;
-            }
-
-            // Los reordeno por golesHechos si ecuentro dos equipos con igual diferencia de goles
-            if (diferenciaDeGolesIguales) {
-                Collections.sort(equipos, new Comparator<Equipo>() {
-                    @Override
-                    public int compare(Equipo e1, Equipo e2) {
-                        int diferenciaDeGolesE1 = e1.getGolesHechos() - e1.getGolesEnContra();
-                        int diferenciaDeGolesE2 = e2.getGolesHechos() - e2.getGolesEnContra();
-
-                        if (diferenciaDeGolesE1 == diferenciaDeGolesE2) {
-                            return e1.getGolesHechos() > e2.getGolesHechos() ? -1 : 1;
-                        }
-                        return 0;
-                    }
-                });
-            }
-        }
-
-        // Este bloque se ejecuta solamente cuando se da click en btn guardar
-        // ...
-        if (guardandoCambios) {
-            // Valido si hay que elegir manualmente segundo o primer puesto
-            boolean necesitaOrdenManual = false;
-
-            // Valido si hay que elegir manualmente segundo puesto
-            boolean segundoYTerceroIgualPuntos = equipos.get(1).getPuntos() == equipos.get(2).getPuntos();
-
-            int dgSegundo = equipos.get(1).getGolesHechos() - equipos.get(1).getGolesEnContra();
-            int dgTercero = equipos.get(2).getGolesHechos() - equipos.get(2).getGolesEnContra();
-            boolean segundoYTerceroIgualDG = dgSegundo == dgTercero;
-
-            boolean segundoYTerceroIgualGolesHechos = equipos.get(1).getGolesHechos() == equipos.get(2).getGolesHechos();
-
-            if (segundoYTerceroIgualPuntos && segundoYTerceroIgualDG && segundoYTerceroIgualGolesHechos) {
-                necesitaOrdenManual = true;
-
-                Object[] opciones = {equipos.get(1).getNombre(), equipos.get(2).getNombre(), equipos.get(3).getNombre()};
-                int seleccion = JOptionPane.showOptionDialog(null, "Debe elegir manualmente el equipo que quedó en 2do puesto", "Selección manual", JOptionPane.DEFAULT_OPTION, JOptionPane.WARNING_MESSAGE, null, opciones, opciones[0]);
-
-                switch (seleccion) {
-                    case 0:
-                        equipos.get(1).setOctavos(true);
-                        break;
-                    case 1:
-                        equipos.get(2).setOctavos(true);
-                        Collections.swap(equipos, 2, 1);
-                        break;
-                    case 2:
-                        equipos.get(3).setOctavos(true);
-                        Collections.swap(equipos, 3, 1);
-                        break;
-                }
-            }
-
-            // Valido si hay que elegir manualmente primer puesto
-            boolean primerYSegundoIgualPuntos = equipos.get(0).getPuntos() == equipos.get(1).getPuntos();
-
-            int dgPrimero = equipos.get(0).getGolesHechos() - equipos.get(0).getGolesEnContra();
-            boolean primerYSegundoIgualDG = dgPrimero == dgSegundo;
-
-            boolean primerYSegundoIgualGolesHechos = equipos.get(0).getGolesHechos() == equipos.get(1).getGolesHechos();
-
-            if (primerYSegundoIgualPuntos && primerYSegundoIgualDG && primerYSegundoIgualGolesHechos) {
-                necesitaOrdenManual = true;
-
-                Object[] opciones = {equipos.get(0).getNombre(), equipos.get(1).getNombre(), equipos.get(2).getNombre()};
-                int seleccion = JOptionPane.showOptionDialog(null, "Debe elegir manualmente el equipo que quedó en 1er puesto", "Selección manual", JOptionPane.DEFAULT_OPTION, JOptionPane.WARNING_MESSAGE, null, opciones, opciones[0]);
-
-                switch (seleccion) {
-                    case 0:
-                        equipos.get(0).setOctavos(true);
-                        break;
-                    case 1:
-                        equipos.get(1).setOctavos(true);
-                        Collections.swap(equipos, 1, 0);
-                        break;
-                    case 2:
-                        equipos.get(2).setOctavos(true);
-                        Collections.swap(equipos, 2, 0);
-                        break;
-                }
-            }
-        }
-
-        // Si hubo que seleccionar de forma manual quienes pasaron hay que impactar ese cambio en la vista de la tabla
-//        if(!equipos.get(1).isOctavos() && equipos.get(2).isOctavos()){
-//            Collections.swap(equipos, 2, 0);
-//        }
-//        if(!equipos.get(0).isOctavos() && equipos.get(1).isOctavos()){
-//            Collections.swap(equipos, 1, 0);
-//        }
-        // Completar equipos de fase de 8vos
-        HashMap<Character, Equipo[]> grupoEquiposPrimerosPuestos = new HashMap<>();
-        Equipo[] equiposParaOctavos = {equipos.get(0), equipos.get(1)};
-
-        grupoEquiposPrimerosPuestos.put(grupo.getLetra(), equiposParaOctavos);
-
-        escribirEquiposEnOctavos(grupoEquiposPrimerosPuestos);
-
-        return equipos;
-    }
-
-    private void escribirEquiposEnOctavos(HashMap<Character, Equipo[]> grupoEquiposPrimerosPuestos) {
-        // Dónde ubicar 1er puesto
-        // Dónde ubicar 2do puesto
-        Character letra = (Character) grupoEquiposPrimerosPuestos.keySet().toArray()[0];
-
-        Partido partidoPrimerPuesto;
-        Partido partidoSegundoPuesto;
-
-        // Seteo el booleano de octavos como verdadero
-        for (Equipo equipo : grupoEquiposPrimerosPuestos.get(letra)) {
-            equipo.setOctavos(true);
-        }
-
-        switch (letra) {
-            case 'a':
-                // Buscar el partido de ID 49
-                partidoPrimerPuesto = fixtureService.obtenerPartidoPorId(49);
-
-                // Buscar el partido de ID 52
-                partidoSegundoPuesto = fixtureService.obtenerPartidoPorId(51);
-
-                partidoPrimerPuesto.setEquipo1(grupoEquiposPrimerosPuestos.get('a')[0]);
-                partidoSegundoPuesto.setEquipo2(grupoEquiposPrimerosPuestos.get('a')[1]);
-
-                break;
-
-            case 'b':
-                // Buscar el partido de ID 52
-                partidoPrimerPuesto = fixtureService.obtenerPartidoPorId(51);
-
-                // Buscar el partido de ID 49
-                partidoSegundoPuesto = fixtureService.obtenerPartidoPorId(49);
-
-                partidoPrimerPuesto.setEquipo1(grupoEquiposPrimerosPuestos.get('b')[0]);
-                partidoSegundoPuesto.setEquipo2(grupoEquiposPrimerosPuestos.get('b')[1]);
-
-                break;
-
-            case 'c':
-                // Buscar el partido de ID 50
-                partidoPrimerPuesto = fixtureService.obtenerPartidoPorId(50);
-
-                // Buscar el partido de ID 52
-                partidoSegundoPuesto = fixtureService.obtenerPartidoPorId(52);
-
-                partidoPrimerPuesto.setEquipo1(grupoEquiposPrimerosPuestos.get('c')[0]);
-                partidoSegundoPuesto.setEquipo2(grupoEquiposPrimerosPuestos.get('c')[1]);
-
-                break;
-
-            case 'd':
-                // Buscar el partido de ID 52
-                partidoPrimerPuesto = fixtureService.obtenerPartidoPorId(52);
-
-                // Buscar el partido de ID 50
-                partidoSegundoPuesto = fixtureService.obtenerPartidoPorId(50);
-
-                partidoPrimerPuesto.setEquipo1(grupoEquiposPrimerosPuestos.get('d')[0]);
-                partidoSegundoPuesto.setEquipo2(grupoEquiposPrimerosPuestos.get('d')[1]);
-
-                break;
-
-            case 'e':
-                // Buscar el partido de ID 53
-                partidoPrimerPuesto = fixtureService.obtenerPartidoPorId(53);
-
-                // Buscar el partido de ID 55
-                partidoSegundoPuesto = fixtureService.obtenerPartidoPorId(55);
-
-                partidoPrimerPuesto.setEquipo1(grupoEquiposPrimerosPuestos.get('e')[0]);
-                partidoSegundoPuesto.setEquipo2(grupoEquiposPrimerosPuestos.get('e')[1]);
-
-                break;
-
-            case 'f':
-                // Buscar el partido de ID 53
-                partidoPrimerPuesto = fixtureService.obtenerPartidoPorId(55);
-
-                // Buscar el partido de ID 55
-                partidoSegundoPuesto = fixtureService.obtenerPartidoPorId(53);
-
-                partidoPrimerPuesto.setEquipo1(grupoEquiposPrimerosPuestos.get('f')[0]);
-                partidoSegundoPuesto.setEquipo2(grupoEquiposPrimerosPuestos.get('f')[1]);
-
-                break;
-
-            case 'g':
-                // Buscar el partido de ID 54
-                partidoPrimerPuesto = fixtureService.obtenerPartidoPorId(54);
-
-                // Buscar el partido de ID 56
-                partidoSegundoPuesto = fixtureService.obtenerPartidoPorId(56);
-
-                partidoPrimerPuesto.setEquipo1(grupoEquiposPrimerosPuestos.get('g')[0]);
-                partidoSegundoPuesto.setEquipo2(grupoEquiposPrimerosPuestos.get('g')[1]);
-
-                break;
-
-            case 'h':
-                // Buscar el partido de ID 56
-                partidoPrimerPuesto = fixtureService.obtenerPartidoPorId(56);
-
-                // Buscar el partido de ID 54
-                partidoSegundoPuesto = fixtureService.obtenerPartidoPorId(54);
-
-                partidoPrimerPuesto.setEquipo1(grupoEquiposPrimerosPuestos.get('h')[0]);
-                partidoSegundoPuesto.setEquipo2(grupoEquiposPrimerosPuestos.get('h')[1]);
-
-                break;
-
-        }
-
-        // Escribo los datos en la pestaña de octavos
-        loadPartidosOctavos();
-
-    }
-
     private void escribirEquiposEnCuartos() {
 
         ArrayList<Partido> partidos = new ArrayList();
@@ -6516,11 +6242,11 @@ public class Ventana extends javax.swing.JFrame {
     }
 
     private void crearYCompletarTablaDePosiciones(Grupo grupo) {
-        HashSet<Equipo> equipos = recuperarDatosDeEquipoDeEquipoRepository(grupo);
+        HashSet<Equipo> equipos = fixtureService.recuperarDatosDeEquipoDeEquipoRepository(grupo);
 
         // Ordeno antes de generar la tabla
-        ArrayList<Equipo> equiposOrdenados = ordenarEquiposYCompletarOctavos(equipos, grupo, false);
-
+        ArrayList<Equipo> equiposOrdenados = fixtureService.ordenarEquiposYCompletarOctavos(equipos, grupo, false);
+        
         ventanaTablaDePosiciones = new TablaDePosiciones(equiposOrdenados);
         ventanaTablaDePosiciones.setDefaultCloseOperation(DISPOSE_ON_CLOSE);
 
@@ -6529,7 +6255,7 @@ public class Ventana extends javax.swing.JFrame {
     }
 
     private void leerGolesDeGrupoYGuardarCambios(Grupo grupo) {
-        HashSet<Equipo> equiposGrupoActualizados = recuperarDatosDeEquipoDeEquipoRepository(grupo);
+        HashSet<Equipo> equiposGrupoActualizados = fixtureService.recuperarDatosDeEquipoDeEquipoRepository(grupo);
 
         // Antes de actualizar los valores para la tabla de cada equipo hay que limpiarlos
         for (Equipo equipo : equiposGrupoActualizados) {
@@ -6615,22 +6341,15 @@ public class Ventana extends javax.swing.JFrame {
 
         // Ordenar los equipos en la tabla de posiciones y guardarlos
         // TODO: Cambiar a true el booleano
-        ArrayList<Equipo> equiposOrdenados = ordenarEquiposYCompletarOctavos(equiposGrupoActualizados, grupo, true);
+        ArrayList<Equipo> equiposOrdenados = fixtureService.ordenarEquiposYCompletarOctavos(equiposGrupoActualizados, grupo, true);
+        
+        // Escribo los datos en la pestaña de octavos
+        loadPartidosOctavos();
+        
         guardarCambios(grupo, equiposOrdenados);
     }
 
-    private HashSet<Equipo> recuperarDatosDeEquipoDeEquipoRepository(Grupo grupo) {
-        HashSet<Equipo> equiposGrupoActualizados = new HashSet();
-        // Para guardar correctamente el dato tengo que trabajar sobre EquipoRepository
-        for (Equipo equipoGrupo : grupo.getEquipos()) {
-            Equipo equipoEncontrado = fixtureService.obtenerEquipo(equipoGrupo);
-
-            // esta lista (HashSet) va a ser la encargada que gestionar los datos de Equipos en el repositorio correspondiente
-            equiposGrupoActualizados.add(equipoEncontrado);
-        }
-
-        return equiposGrupoActualizados;
-    }
+    
 
     private void guardarCambios(Grupo grupo, ArrayList<Equipo> equiposGrupoActualizados) throws HeadlessException {
         try {
